added getTotal to financialEventControllers
